requirements.txt
numpy==1.18.5
opencv-python==4.2.0.34
pandas==1.0.5
Pillow==7.1.2
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
